lib/string: use the biggest allowed size in memcpy
The generic memcpy is defined as a one by one byte copy. This is always
safe to use, but it's slower than a 4 byte or even 8 byte copy.
The problem with bigger size copy is that it needs data to be properly
aligned.
Write a generic memcpy which copies the data using the maximum aligned
size allowed by the pointer alignment.

First copy data one byte at time until the destination pointer is
aligned to a long boundary, then copy using the biggest size which won't
cause unaligned accesses.

Signed-off-by: Matteo Croce <mcroce@microsoft.com>