riscv: optimized memset
The generic memset is defined as a byte at one set. This is always safe
to use but it's slower than a 4 byte or even 8 byte set.
The problem with bigger size set is that it needs data to be properly
aligned.
Write a generic memset which fills the data one by one until the
destination is aligned, then fills using the largest size allowed,
and finally fills the remaining data one byte at time.

Signed-off-by: Matteo Croce <mcroce@microsoft.com>